api/requirements.txt
fastapi==0.85.1
uvicorn==0.22.0
httpx==0.24.0
python-dotenv==1.0.0
pypdf==3.10.0
